+++ test/S3ErgodicMovesTest.cpp
@@ -16,8 +16,8 @@ SCENARIO("Perform ergodic moves upon S3 Triangulations", "[moves]")
 {
   GIVEN("A 3D 2-sphere foliated triangulation.")
   {
-    constexpr auto     simplices  = static_cast<std::intmax_t>(32000);
-    constexpr auto     timeslices = static_cast<std::intmax_t>(12);
+    constexpr auto     simplices  = static_cast<std::int_fast32_t>(32000);
+    constexpr auto     timeslices = static_cast<std::int_fast32_t>(12);
     SimplicialManifold universe(simplices, timeslices);
     Move_tracker       attempted_moves{{0, 0, 0, 0, 0}};
     // Verify triangulation
